Convert StatsPolicyConfig to proto bytes
Convert the existing StatsPolicyConfig object to a wire-encoded proto
byte array. Based on an email thread with API council, this is an
appropriate use case of a byte array.

Also, responded to API council request for documentation for what
happens when you get invalid params on
setRestrictedMetricsChangedOperation.

Bug: 271056938
Bug: 269604773
Test: m
Test: atest statsd_test
Ignore-AOSP-First: Android U API & feature
Change-Id: Ifbd7b63c493d8a25d1d25317db374c33de7f76e8