Fix libstatspull.map.txt function names
A recent CL (ag/21489712) introduced three new functions to
libstatspull. The names of these functions were updated during the
review of that CL. In the end the libstatspull.map.txt file ended
up using the old names:
- AStatsManager_subscribeToAtomEvents
- AStatsManager_unsubscribeFromAtomEvents
- AStatsManager_flushAtomEvents
Where as the .h and .cc files use the newer names:
- AStatsManager_addSubscription
- AStatsManager_removeSubscription
- AStatsManager_flushSubscription

Update libstatspull.map.txt to use the new names.

Ignore-AOSP-First: Adding #apex API that cannot be added to AOSP first
as it automerges into tm-mainline-prod
Test: Build with WIP CL that links libstatspull and uses the new APIs
Bug: 268661096

Change-Id: Ic6139dba0fd24c232417aa872a3fbbdc49945a82