Log whether metric hits dimension guardrail
For duration and value metrics, the state of dimension_guardrail_hit
will be true for all reports after the guardrail is hit until the device
is rebooted.

For count and gauge metrics, the state of dimension_guardrail_hit is
logged on a per report basis, hitting the guardrail for one report is
not expected to impact the data of other reports.

Test: atest statsd_test
Bug: 292645997
Ignore-AOSP-First: merge conflicts in aosp
Change-Id: I023e2e8e8a82ca12f2af2eda02c6ba1c2504bf4e
Merged-In: I023e2e8e8a82ca12f2af2eda02c6ba1c2504bf4e