Don't reset condition when event is late
When an event is late, the current bucket is dropped since data could be
attributed to the wrong condition. However, the condition was also reset
to unknown, with the idea that a later event from the future could have
already arrived, and this late event would set the condition to an
incorrect value.

However, statsd already relies on the condition events being logged in
order in many other places (condition trackers, other metrics, esp
duration metric, and it does not make sense to use the bucket boundary
as a threshold for declaring the condition incorrect.

Test: atest statsd_test
Bug: 296593839
Change-Id: I66895d1834c3591a3eec45397c61927de3cc7fb2