Remove cts tag of MainlineCTS StatsD test modules.
MCTS modules will be removed from android-cts.zip and packaged into android-mcts.zip. And it will be downloaded automatically when run cts via cts-tradefed in android-cts.zip.

Ignore-AOSP-First: The dynamic downloading logic was only added in internal main branch. If this change is submitted in AOSP, the MCTS modules will not be executed in CTS from AOSP main branch.  So we decide to submit the change in internal main. After public release, the dynamic downloading change and this change will be snapped to aosp main together.

Bug: 316436147
Test: abtd
Change-Id: Ice7c6820afda5f841934730a89c15585d64a40bb