Use NumericValue based on std::variant instead of Value in NumericValueMetricProducer
Value from FieldValue is too heavy for ValueMetric as it has support for extra types like strings and
byte arrays that we don't need and take up extra memory on the stack.
In the future, NumericValue std::variant can be extended to include
histograms without affecting other usages of Value.

Bug: 185796344
Test: statsd_test + ValueMetric CTS
Change-Id: Ib5e381baba19b89b72551b7493bffe80afe2a80f