Add statsd-validation for client-aggregated histograms.
Add a new entry in the binning_strategy oneof in HistogramBinConfig to
indicate the binning information is not provided by the config.
value_field in ValueMetric now requires a repeated field with Position
ALL if the corresponding aggregation_type is HISTOGRAM and the
binning_strategy in HistogramBinConfig is set to client_aggregated_bins.
If ValueMetric has a histogram, value_direction can only be INCREASING.

Bug: 352222583
Test: statsd_test
Change-Id: I94817ed521a5f0af483d0fd399e5f15c4525b046