Implement client-aggregated histograms in statsd
Histograms are represented as repeated int32 fields inside the atom.
HistogramBinConfig needs to specify client_aggregated_bins as the
binning_strategy.
In the ValueMetric, value_field needs a matcher that points to the
repeated int32 atom field with Position set to ALL.

Statsd takes element-wise diffs of the repeated int32 values for diffed
metrics.

Bug: 352222583
Test: m statsd
Test: statsd_test
Change-Id: I54567a13350d03204fb9e1f5b2982c6be9e664fd