Add ability to remove unused uids from the uid map
Controlled by a config flag, and also add an allowlist of packages that
will always be included, regardless of whether they are used or are
system uid.

Used uids are tracked when the dump report happens, similar to the
string set.

TODO: add tests. Will be done in this cl, just
sending out now to expedite review.

Test: atest statsd_test
Bug: 365698996
Change-Id: I5a85dd1660acf1d43c5991c2a4e65a3d54045e9c