Add ability to specify uid fields in the metric
This allows metrics to specify uid fields in case the atom is missing
them. This is not meant to be commonly used, as annotating the atom is
the preferred way of marking uid fields. However, if the atom is not
properly annotated, this is a backup option that allows statsd to track
used uid fields properly.

Note that this does not work for any other use cases of the uid
annotation, such as matching on package names or isolated to host uid
mapping.

Test: atest statsd_test
Bug: 374234484
Change-Id: Ie4c24f5044245a330eb4494a7fdda556b1028f0b